+++ src/main/java/iesmb/pp3/planillaProfesores/service/jpa/PuntajeActividadServiceImpl.java
@@ -99,10 +99,10 @@ public class PuntajeActividadServiceImpl implements IPuntajeActividadService{
 		nuevoPuntaje.setPuntaje(0); // Otra lógica si es necesario
 		return nuevoPuntaje;
 	}
-	public int obtenerTotalPuntosPorProfesor(Profesor profesor) {
+	public float obtenerTotalPuntosPorProfesor(Profesor profesor) {
 		List<PuntajeActividad> puntajesPorProfesor = puntajeActividadRepository.findByProfesor(profesor);
 
-		int totalPuntos = 0;
+		float totalPuntos = 0;
 
 		for (PuntajeActividad puntaje : puntajesPorProfesor) {
 			totalPuntos += puntaje.getPuntaje();
